linuxkernel: Refactor the code to allow addition of ARM images
We need to detect the architecture of the binary so we can handle it
accordingly; to do so, we need to abstract the detection from the
version detection code.

This commit prepares the basis for it, as it adds a new method to be
used to detect the Linux kernel image type and consequently the
architecture.

Signed-off-by: Otavio Salvador <otavio@ossystems.com.br>